ide: Folding
